Refactor: 로딩 시 버튼 클릭 막고 useScrapMutation 훅 제작
